requirements.txt
pandas
astroquery
Astropy
scipy
Pillow
matplotlib
photutils